+++ React/Shared/Components/Course/LS/LSMap.js
@@ -24,10 +24,24 @@
 /**
  * Following learning materials are needed in both verbal and visual formats.
  * 
- * Overview 
- * Definitions
- * Contents
- * Activities
- * Summary
+ * Overview - 2
+ *  Verbal -    https://www.tutorialspoint.com/java/index.htm
+ *  Visual -    https://youtu.be/NcmDpIFVJN4
+ * Definitions - 4
+ *  Verbal - Detailed - https://www.tutorialspoint.com/java/java_basic_datatypes.htm
+ *  Verbal - OverView - https://www.tutorialspoint.com/java/java_basic_datatypes.htm ( Needes To Be Cutshorted )
+ *  Visual - Detailed - https://www.javatpoint.com/java-data-types
+ *  Visual - Overview - https://www.scientecheasy.com/2018/06/data-types-in-java-primitive-example.html 
+ * Contents - 4
+ *  Verbal - Detailed - https://www.tutorialspoint.com/java/java_basic_datatypes.htm
+ *  Verbal - Overview - https://www.tutorialspoint.com/java/java_basic_datatypes.htm ( Needes To Be Cutshorted ) 
+ *  Visual - Detailed - https://youtu.be/bqPIWlnjWbA
+ *  Visual - Overview - https://www.youtube.com/watch?v=snIUtdg0K30
+ * Activities - 2
+ *  Verbal - https://beginnersbook.com/2017/08/data-types-in-java/
+ *  Visual - https://www.youtube.com/watch?v=vhV97hyV0fc 
+ * Summary - 2
+ *  Verbal - https://beginnersbook.com/2017/08/data-types-in-java/ 
+ *  Visual - https://i1.wp.com/www.guffo.in/wp-content/uploads/2018/03/3.png?ssl=1 
  * 
  */
